+++ apps/www/src/registry/default/plate-ui/blockquote-element.tsx
@@ -3,18 +3,19 @@
 import React from 'react';
 
 import { cn, withRef } from '@udecode/cn';
-import { PlateElement } from '@udecode/plate-common/react';
+
+import { PlateElement } from './plate-element';
 
 export const BlockquoteElement = withRef<typeof PlateElement>(
   ({ children, className, ...props }, ref) => {
     return (
       <PlateElement
         ref={ref}
-        asChild
+        as="blockquote"
         className={cn('my-1 border-l-2 pl-6 italic', className)}
         {...props}
       >
-        <blockquote>{children}</blockquote>
+        {children}
       </PlateElement>
     );
   }