+++ apps/www/src/registry/default/plate-ui/date-element.tsx
@@ -2,9 +2,10 @@
 
 import { cn, withRef } from '@udecode/cn';
 import { setNodes } from '@udecode/plate-common';
-import { PlateElement, findNodePath } from '@udecode/plate-common/react';
+import { findNodePath } from '@udecode/plate-common/react';
 
 import { Calendar } from './calendar';
+import { PlateElement } from './plate-element';
 import { Popover, PopoverContent, PopoverTrigger } from './popover';
 
 export const DateElement = withRef<typeof PlateElement>(