+++ apps/www/src/registry/default/plate-ui/emoji-input-element.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo, useState } from 'react';
 
 import { withRef } from '@udecode/cn';
-import { PlateElement } from '@udecode/plate-common/react';
+import { PlateElement } from './plate-element';
 import { EmojiInlineIndexSearch, insertEmoji } from '@udecode/plate-emoji';
 
 import { useDebounce } from '@/registry/default/hooks/use-debounce';