+++ apps/www/src/registry/default/plate-ui/heading-element.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 
 import { withRef, withVariants } from '@udecode/cn';
-import { PlateElement } from '@udecode/plate-common/react';
 import { cva } from 'class-variance-authority';
+
+import { PlateElement } from './plate-element';
 
 const headingVariants = cva('relative mb-1', {
   variants: {
@@ -36,17 +37,15 @@ export const HeadingElement = withRef<typeof HeadingElementVariants>(
   ({ children, isFirstBlock, variant = 'h1', ...props }, ref) => {
     const { editor, element } = props;
 
-    const Element = variant!;
-
     return (
       <HeadingElementVariants
         ref={ref}
-        asChild
+        as={variant!}
         variant={variant}
         isFirstBlock={element === editor.children[0]}
         {...props}
       >
-        <Element>{children}</Element>
+        {children}
       </HeadingElementVariants>
     );
   }