+++ apps/www/src/registry/default/plate-ui/image-element.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 
 import { cn, withRef } from '@udecode/cn';
-import { PlateElement, withHOC } from '@udecode/plate-common/react';
+import { withHOC } from '@udecode/plate-common/react';
 import { Image, ImagePlugin, useMediaState } from '@udecode/plate-media/react';
 import { ResizableProvider, useResizableStore } from '@udecode/plate-resizable';
 
 import { Caption, CaptionTextarea } from './caption';
 import { MediaPopover } from './media-popover';
+import { PlateElement } from './plate-element';
 import {
   Resizable,
   ResizeHandle,