+++ apps/www/src/registry/default/plate-ui/mention-input-element.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 
 import { cn, withRef } from '@udecode/cn';
-import { PlateElement } from '@udecode/plate-common/react';
+import { PlateElement } from './plate-element';
 import { getMentionOnSelectItem } from '@udecode/plate-mention';
 
 import {