+++ apps/www/src/registry/default/plate-ui/paragraph-element.tsx
@@ -1,4 +1,20 @@
-import { withCn } from '@udecode/cn';
-import { PlateElement } from '@udecode/plate-common/react';
+import React from 'react';
 
-export const ParagraphElement = withCn(PlateElement, 'm-0 px-0 py-1');
+import { cn } from '@udecode/cn';
+import { withRef } from '@udecode/plate-common/react';
+
+import { PlateElement } from './plate-element';
+
+export const ParagraphElement = withRef<typeof PlateElement>(
+  ({ children, className, ...props }, ref) => {
+    return (
+      <PlateElement
+        ref={ref}
+        className={cn('m-0 px-0 py-1', className)}
+        {...props}
+      >
+        {children}
+      </PlateElement>
+    );
+  }
+);