+++ packages/react-utils/src/MemoizedChildren.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const MemoizedChildren = React.memo(
+  ({ children }: { children: React.ReactNode }) => {
+    return <>{children}</>;
+  }
+);