+++ packages/find-replace/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/plate-find-replace
 
+## 39.0.0
+
 ## 38.0.0
 
 ## 37.0.0