+++ packages/kbd/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/plate-kbd
 
+## 39.0.0
+
 ## 38.0.1
 
 ### Patch Changes